Stock_market_analysis/requirements.txt
pandas
matplotlib
seaborn
mysql-connector-python
python-dotenv
